# Updated is_ref to make unmarshal_response ~30% faster