--- a/tox.ini
+++ b/tox.ini
@@ -5,7 +5,7 @@ envlist = py27, py35, py36, pre-commit
 deps =
     -rrequirements-dev.txt
 commands =
-    python -m pytest --capture=no {posargs:tests} --benchmark-min-rounds=25
+    python -m pytest --capture=no {posargs:tests} --benchmark-min-rounds=10
 
 [testenv:cover]
 basepython = /usr/bin/python2.7