--- a/tox.ini
+++ b/tox.ini
@@ -5,7 +5,13 @@ envlist = py27, py35, py36, pre-commit
 deps =
     -rrequirements-dev.txt
 commands =
-    python -m pytest --capture=no {posargs:tests} --benchmark-min-rounds=10
+    python -m pytest --capture=no {posargs:tests} 
+
+[testenv:benchmark]
+deps =
+    -rrequirements-dev.txt
+commands =
+    python -m pytest --capture=no {posargs:tests/profiling} --benchmark-min-rounds=15
 
 [testenv:cover]
 basepython = /usr/bin/python2.7
@@ -34,4 +40,4 @@ deps =
 setenv =
     LC_CTYPE=en_US.UTF-8
 commands =
-    pre-commit {posargs:run --all-files}
+    pre-commit run --all-files