@ClareCat 2017-08-22
The benchmarking tests were making testing much slower (especially during dev) so I adjusted the make target for 'test' to add args that will ignore the benchmarking tests and added a 'benchmark' target. Travis will still run all of the tests because it just runs tox without any of the ignore posargs.

Noticed during this that the way the tox environments were setup it was impossible to pass in any posargs for pytest that wouldn't crash in pre-commit so opted to remove posargs from the pre-commit testenv.